Refactor requests to allow for non Normandy requests
